// JcCar-client/src/app/shared/model/enumaration/status.alert.ts
export enum StatusAlertEnum {
    REGISTRADO = 1,
    AGUARDANDO_ENVIO = 2,
    AGUARDANDO_RESPOSTA = 3,
    AVALIADO_ABRANGENTE = 4,
    AVALIADO_NAO_ABRANGENTE = 5,
    PLANO_ACAO = 6,
    PLANO_ACAO_EM_AVALIACAO = 7,
    PLANO_ACAO_APROVADO = 8,
    PLANO_ACAO_REPROVADO = 9,
    ALERTA_EM_AVALIACAO = 10,
    ALERTA_REPROVADO = 11,
    ALERTA_APROVADO = 12
}